% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bargain_foc_calibrate.R
\name{bargain_foc_calibrate}
\alias{bargain_foc_calibrate}
\title{Nash bargaining calibration}
\usage{
bargain_foc_calibrate(
  param,
  ownership,
  price,
  shares,
  cost,
  weight,
  lambda,
  includeMUI = TRUE
)
}
\arguments{
\item{price}{Price}

\item{cost}{Marginal costs for each product}

\item{lambda}{Bargaining power of the buyer}

\item{includeMUI}{logical; TO BE ADDED}

\item{own}{Ownership matrix}

\item{alpha}{Price coefficient}

\item{delta}{Mean values}
}
\value{
The first-order conditions
}
\description{
Nash bargaining calibration
}
\details{
This function calculate the first-order conditions from a Bertrand
price-setting model of competition
}
\examples{
TO BE ADDED.

}
