% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bargain_foc_vert_sim.R
\name{bargain_foc_vert_sim}
\alias{bargain_foc_vert_sim}
\title{Nash Product FOCs for use in vertical model with simultaneous timing}
\usage{
bargain_foc_vert_sim(
  price_w,
  own_down,
  own_up,
  alpha,
  delta,
  cost_w,
  cost_r,
  lambda,
  price_r,
  sumFOC = TRUE
)
}
\arguments{
\item{price_w}{Upstream or wholesale prices}

\item{own_down}{Ownership matrix for downstream firms}

\item{own_up}{Ownership matrix for upstream firms}

\item{alpha}{Price coefficient}

\item{delta}{Mean values}

\item{cost_w}{Marginal costs for upstream firm for each product}

\item{cost_r}{Marginal costs for downstream firm for each product}

\item{lambda}{Bargaining power of the buyer/retailer}

\item{price_r}{Retail prices starting values}

\item{sumFOC}{logical; if true, returns the sum of first-order conditions, if
false, returns vector of FOC's for each product}
}
\value{
The first-order conditions
}
\description{
Nash Product FOCs for use in vertical model with simultaneous timing
}
\details{
This function calculate the first-order conditions from a Nash
bargaining model. For use in a vertical supply chain with simultaneous timing
}
\examples{
bargain_foc_vert_sim(price_w = c(1.6, 1.6, 1.6, 1.6, 1.6, 1.6),
own_down = paste0("R",rep(c(1,2,3),each=2)),
own_up = paste0("W",rep(c(1,2),3)),
alpha = -0.9, delta = c(0.2, 0.3, 0.9, 1.0, 0.8, 0.9),
cost_w = rep(.2, times = 6),
cost_r = rep(.1, times = 6),
lambda = 0.5,
price_r = c(2.9, 2.9, 3.0, 3.0, 3.0, 3.0))

}
