% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bargain_NP_vert_seq.R
\name{bargain_NP_vert_seq_gnl}
\alias{bargain_NP_vert_seq_gnl}
\title{Nash Product for use in vertical model with sequential timing}
\usage{
bargain_NP_vert_seq_gnl(
  w_start,
  product_max,
  price_w,
  own_down,
  own_up,
  alpha,
  delta,
  cost_w,
  cost_r,
  lambda,
  p_R0,
  sigma,
  nest_allocation,
  mu,
  showAll = FALSE
)
}
\arguments{
\item{w_start}{Initial wholesale price for product for which to maximize Nash Product}

\item{product_max}{Index of product for which to maximize Nash Product}

\item{price_w}{Upstream or wholesale prices}

\item{own_down}{Ownership matrix for downstream firms}

\item{own_up}{Ownership matrix for upstream firms}

\item{alpha}{Price coefficient}

\item{delta}{Mean values}

\item{cost_w}{Marginal costs for upstream firm for each product}

\item{cost_r}{Marginal costs for downstream firm for each product}

\item{lambda}{Bargaining power of the buyer/retailer}

\item{p_R0}{Retail prices starting values}

\item{sigma}{Contract type; value between 0 and 1 where 0 is linear price,
1 is two-part tariff}

\item{nest_allocation}{For generalized nested logit demand, a J-by-K matrix
where each element (j,k) designates the membership of good j in nest k. Rows
should sum to 1.}

\item{mu}{Nesting parameters for each nest}

\item{showAll}{logical; if true, returns gains for trade for every product}
}
\value{
The first-order conditions
}
\description{
Nash Product for use in vertical model with sequential timing
}
\details{
This function calculate the Nash Product from a Nash bargaining
model. This version allows for nested logit with overlapping nests.
}
\examples{

own_down <- paste0("R",rep(c(1,2,3),each=2))
own_up <- paste0("W",rep(c(1,2),3))

B1 <- 1 * matrix( c(own_down == "R1",
                    own_down == "R2",
                    own_down == "R3",
                    own_up == "W1",
                    own_up == "W2"),
                  ncol = 5, nrow = 6)
a1 <- B1 * 0.5
mu1 <- rep(1.0,5)

bargain_NP_vert_seq_gnl(w_start = 1.5, product_max = 1,
                        price_w = c(1.6, 1.6, 1.6, 1.6, 1.6, 1.6),
                        own_down = own_down,
                        own_up = own_up,
                        alpha = -0.9, delta = c(0.2, 0.3, 0.9, 1.0, 0.8, 0.9),
                        cost_w = rep(.2, times = 6),
                        cost_r = rep(.1, times = 6),
                        lambda = 0.5, sigma = 0,
                        p_R0 = c(2.9, 2.9, 3.0, 3.0, 3.0, 3.0),
                        nest_allocation = a1, mu = mu1)


}
