% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bargain_vert_seq_calibrate.R
\name{bargain_vert_seq_calibrate}
\alias{bargain_vert_seq_calibrate}
\title{Calibrate costs for Nash bargain in vertical model with sequential timing}
\usage{
bargain_vert_seq_calibrate(
  c_w_val,
  price_w,
  own_down,
  own_up,
  alpha,
  delta,
  lambda,
  cost_r,
  price_r,
  sigma,
  setTol = 0.01,
  setMaxIter = 500,
  showAll = FALSE,
  symmetricCosts = FALSE
)
}
\arguments{
\item{c_w_val}{Wholesale/upstream costs}

\item{price_w}{Upstream or wholesale prices}

\item{own_down}{Ownership matrix for downstream firms}

\item{own_up}{Ownership matrix for upstream firms}

\item{alpha}{Price coefficient}

\item{delta}{Mean values}

\item{lambda}{Bargaining power of the buyer/retailer}

\item{cost_r}{Marginal costs for downstream firm for each product}

\item{price_r}{Retail prices}

\item{sigma}{Contract type; value between 0 and 1 where 0 is linear price,
1 is two-part tariff}

\item{setTol}{tolerance for convergence}

\item{setMaxIter}{Maximum iterations for convergence loop}

\item{showAll}{logical; if true, returns gains for trade for every product}

\item{symmetricCosts}{Constrain costs to be symmetric}
}
\value{
The value of objective function
}
\description{
Calibrate costs for Nash bargain in vertical model with sequential timing
}
\details{
This function can be used to calibrate the costs in a Nash bargain
which is the upstream market of a vertical supply chain. Assumes logit demand.
}
\examples{
TO BE ADDED.

}
