% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bargain_vert_sim_calibrate.R
\name{bargain_vert_sim_calibrate}
\alias{bargain_vert_sim_calibrate}
\title{Calibrate lambda for Nash bargain in vertical model with simultaneous timing}
\usage{
bargain_vert_sim_calibrate(
  lambda,
  price_w,
  own_down,
  own_up,
  alpha,
  delta,
  cost_w,
  cost_r,
  price_r
)
}
\arguments{
\item{lambda}{Bargaining power of the buyer/retailer}

\item{price_w}{Upstream or wholesale prices}

\item{own_down}{Ownership matrix for downstream firms}

\item{own_up}{Ownership matrix for upstream firms}

\item{alpha}{Price coefficient}

\item{delta}{Mean values}

\item{cost_w}{Marginal costs for upstream firm for each product}

\item{cost_r}{Marginal costs for downstream firm for each product}

\item{price_r}{Retail prices}
}
\value{
The value of objective function
}
\description{
Calibrate lambda for Nash bargain in vertical model with simultaneous timing
}
\details{
This function can be used to calibrate the bargaining parameter
in a Nash bargain which is the upstream market of a vertical supply chain.
Assumes logit demand.
}
\examples{
TO BE ADDED.

}
