% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertrand_calibrate.R
\name{bertrand_calibrate}
\alias{bertrand_calibrate}
\title{Bertrand model calibration}
\usage{
bertrand_calibrate(
  param,
  own,
  price,
  shares,
  cost,
  weight,
  returnOutcomes = FALSE
)
}
\arguments{
\item{param}{Price coefficient alpha parameter to calibrate}

\item{own}{Ownership matrix}

\item{price}{Price}

\item{shares}{Observed market shares}

\item{cost}{Marginal costs for each product}

\item{weight}{Weighting matrix}

\item{returnOutcomes}{logical; should equilibrium objects be returned (mean
value parameter, prices, shares, costs) as a list.}
}
\value{
Distance between observed values and model predicted values for
prices and shares
}
\description{
Bertrand model calibration
}
\details{
This function calculate the first-order conditions from a Bertrand
price-setting model of competition. This function is only for standard logit
demand. For nested logit or generalized nested logit, see
bertrand_calibrate_gnl().
}
\examples{
alpha  <- -0.9
delta <- c(.81,.93,.82)
c_j <- c(.05,.31,.30)

own_pre = diag(3)

p0 <- c_j*1.1

share1 <- (exp(delta + alpha*p0))/(1+sum(exp(delta + alpha*p0)))
x00 <- c(-1)
wt_matrix <- diag(c(1,1,1,1000,1000,1000))

bertrand_calibrate( param = x00,
                    own = own_pre, price = p0,
                    shares = share1, cost  = c_j,
                    weight = wt_matrix)

}
