% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertrand_calibrate_gnl.R
\name{bertrand_calibrate_gnl}
\alias{bertrand_calibrate_gnl}
\title{Calibrate Bertrand model with GNL demand}
\usage{
bertrand_calibrate_gnl(
  param,
  own,
  price,
  shares,
  cost,
  weight = c(1, 1, 1, 1),
  nest_allocation,
  div_matrix,
  mu_constraint_matrix = NA,
  div_calc_marginal = TRUE,
  optimizer = "BBoptim",
  optimizer_c = "optim",
  returnOutcomes = FALSE,
  maxitval = 1500,
  maxitval_c = 1500,
  fast_version = FALSE
)
}
\arguments{
\item{param}{Vector of demand parameters (alpha,mu)}

\item{own}{Ownership matrix}

\item{price}{Observed prices}

\item{shares}{Observed market shares}

\item{cost}{Marginal costs for each product}

\item{weight}{Vector of length four with weights given to prices, shares,
diversions, and costs, respectively. Default is c(1,1,1,1).}

\item{nest_allocation}{For generalized nested logit demand, a J-by-K matrix
where each element (j,k) designates the membership of good j in nest k. Rows
should sum to 1.}

\item{div_matrix}{A matrix of observed diversions from product in row j to
product in column k.}

\item{mu_constraint_matrix}{is a (K-by-K') matrix indicating which nesting
parameters are constrained to be equal to each other, where K is the
number of nests and K' is the number of freely varying nesting
parameters. mu_full = mu_constraint_matrix \%*\% mu_prime. Where mu_full
is a vector of length K of the nesting parameter value for each nest,
and mu_prime is a vector of length K' of parameters to be calculated.
It must be the case that K is greater than K'.}

\item{div_calc_marginal}{is a logical if function should match to marginal
diversions (if TRUE) or second choice diversions (if FALSE). Default
to TRUE.}

\item{optimizer}{Which optimization routine should be used to find
equilibrium prices, either BBoptim or multiroot}

\item{optimizer_c}{Which optimization routine should be used to calibrate
costs, either BBoptim or optim}

\item{returnOutcomes}{logical; should equilibrium objects be returned (mean
value parameter, prices, shares, costs) as a list.}

\item{maxitval}{Max iterations during price optimization from foc}

\item{maxitval_c}{Max iterations during cost optimization from foc}

\item{fast_version}{logical; if True, uses only first-order conditions for
products that have non-missing costs. If False, uses the model to predict costs
for all products, and then uses all first-order conditions, which takes longer
to calibrate. Default is False.}
}
\value{
Difference between model predicted and observed values of
prices, shares, and diversions.
}
\description{
Calibrate Bertrand model with GNL demand
}
\details{
This function calibrates a Bertrand model with generalized nested
logit (GNL) demand
}
\examples{
nest1 <- matrix( c(1, 0, 0, 0, 1, 1), ncol = 2, nrow = 3)
divmat <- matrix( c(0, .4, .4, .4, 0, .4, .4, .4, 0), ncol = 3, nrow = 3)

bertrand_calibrate_gnl(param = c(-0.9, 1, 1),
                       own = diag(3),
                       price = c(.05, .34, .33),
                       shares = c( 0.31, 0.27, 0.25),
                       cost = c(.05,.31,.30),
                       weight = c(1,1,1,1),
                       nest_allocation = nest1, div_matrix = divmat)

}
