% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertrand_calibrate_mu.R
\name{bertrand_calibrate_mu}
\alias{bertrand_calibrate_mu}
\title{Calibrate Bertrand model with GNL demand}
\usage{
bertrand_calibrate_mu(
  param,
  own,
  alpha,
  price,
  shares,
  cost,
  weight,
  nest_allocation,
  div_matrix,
  mu_constraint_matrix = NA,
  div_calc_marginal = TRUE,
  optimizer = "BBoptim",
  useOldWeight = FALSE,
  returnOutcomes = FALSE
)
}
\arguments{
\item{param}{Vector of nesting parameters (mu)}

\item{own}{Ownership matrix}

\item{alpha}{Value of price coefficient, to be held fixed during calibration}

\item{price}{Observed prices}

\item{cost}{Marginal costs for each product}

\item{weight}{Vector of weights given to prices, shares, diversions, respectively}

\item{nest_allocation}{For generalized nested logit demand, a J-by-K matrix
where each element (j,k) designates the membership of good j in nest k. Rows
should sum to 1.}

\item{div_matrix}{A matrix of observed diversions from product in row j to
product in column k.}

\item{mu_constraint_matrix}{is a (K-by-K') matrix indicating which nesting
parameters are constrained to be equal to each other, where K is the
number of nests and K' is the number of freely varying nesting
parameters. mu_full = mu_constraint_matrix \%*\% mu_prime. Where mu_full
is a vector of length K of the nesting parameter value for each nest,
and mu_prime is a vector of length K' of parameters to be calculated.
It must be the case that K is greater than K'.}

\item{div_calc_marginal}{is a logical if function should match to marginal
diversions (if TRUE) or second choice diversions (if FALSE). Default
to TRUE.}

\item{optimizer}{Which optimization routine should be used to find
equilibrium prices, either BBoptim or multiroot}

\item{returnOutcomes}{logical; should equilibrium objects be returned (mean
value parameter, prices, shares, costs) as a list.}

\item{share}{Observed market shares}
}
\value{
Difference between model predicted and observed values of
prices, shares, and diversions.
}
\description{
Calibrate Bertrand model with GNL demand
}
\details{
This function calibrates a Bertrand model with generalized nested
logit (GNL) demand, using only first-order conditions that are available,
i.e. first-order conditions for products that have non-missing costs.
}
\examples{
TO BE ADDED.
}
