% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertrand_foc.R
\name{bertrand_foc}
\alias{bertrand_foc}
\title{Bertrand first-order conditions}
\usage{
bertrand_foc(
  price,
  own,
  alpha,
  delta,
  cost,
  nest_allocation = NA,
  mu = NA,
  sumFOC = FALSE
)
}
\arguments{
\item{price}{Price}

\item{own}{Ownership matrix}

\item{alpha}{Price coefficient}

\item{delta}{Mean values}

\item{cost}{Marginal costs for each product}

\item{nest_allocation}{For generalized nested logit demand, a J-by-K matrix
where each element (j,k) designates the membership of good j in nest k. Rows
should sum to 1.}

\item{mu}{Nesting parameters for each nest}

\item{sumFOC}{logical; whether to return the sum of squares of
the first-order conditions. Defaults to FALSE, in which case it returns each
product first-order condition as a vector.}
}
\value{
The first-order conditions
}
\description{
Bertrand first-order conditions
}
\details{
This function calculate the first-order conditions from a Bertrand
price-setting model of competition
}
\examples{
TO BE ADDED.

}
