% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertrand_calibrate.R
\name{bertrand_foc_calibrate}
\alias{bertrand_foc_calibrate}
\title{Bertrand model calibration}
\usage{
bertrand_foc_calibrate(param, own, price, shares, cost, weight)
}
\arguments{
\item{own}{Ownership matrix}

\item{price}{Price}

\item{cost}{Marginal costs for each product}

\item{alpha}{Price coefficient}

\item{delta}{Mean values}

\item{a_jk}{For generalized nested logit demand}

\item{B}{For generalized nested logit demand}

\item{mu}{Nesting parameters for each nest}

\item{sumFOC}{logical; whether to return the sum of squares of
the first-order conditions. Defaults to FALSE, in which case it returns each
product first-order condition as a vector.}
}
\value{
The first-order conditions
}
\description{
Bertrand model calibration
}
\details{
This function calculate the first-order conditions from a Bertrand
price-setting model of competition
}
\examples{
TO BE ADDED.

}
