% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertrand_foc_logit.R
\name{bertrand_foc_logit}
\alias{bertrand_foc_logit}
\title{Bertrand first-order conditions}
\usage{
bertrand_foc_logit(price, own, alpha, delta, cost, sumFOC = FALSE)
}
\arguments{
\item{price}{Price}

\item{own}{Ownership matrix}

\item{alpha}{Price coefficient}

\item{delta}{Mean values}

\item{cost}{Marginal costs for each product}

\item{sumFOC}{logical; whether to return the sum of squares of
the first-order conditions. Defaults to FALSE, in which case it returns each
product first-order condition as a vector.}
}
\value{
The first-order conditions
}
\description{
Bertrand first-order conditions
}
\details{
This function calculate the first-order conditions from a Bertrand
price-setting model of competition
}
\examples{
TO BE ADDED.
}
