% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertrand_foc_vert.R
\name{bertrand_foc_vert}
\alias{bertrand_foc_vert}
\title{Downstream Bertrand first-order conditions}
\usage{
bertrand_foc_vert(
  price_r,
  own_down,
  own_up,
  alpha,
  delta,
  cost_r,
  price_w,
  cost_w,
  sumFOC = FALSE
)
}
\arguments{
\item{price_r}{Downstream or retail prices}

\item{own_down}{Ownership matrix for downstream firms}

\item{own_up}{Ownership matrix for upstream firms}

\item{alpha}{Price coefficient}

\item{delta}{Mean values}

\item{cost_r}{Marginal costs for downstream firm for each product}

\item{price_w}{Upstream or wholesale prices, treated as costs by downstream
firms}

\item{cost_w}{Marginal costs for upstream firm for each product}

\item{sumFOC}{logical; whether to return the sum of squares of
the first-order conditions. Defaults to FALSE, in which case it returns each
product first-order condition as a vector.}
}
\value{
The first-order conditions
}
\description{
Downstream Bertrand first-order conditions
}
\details{
This function calculates the first-order conditions from a Bertrand
price-setting model of competition as the downstream market of a vertical
supply chain. Assumes logit demand.
}
