% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertrand_vert_calibrate.R
\name{bertrand_vert_calibrate}
\alias{bertrand_vert_calibrate}
\title{Calibrate Bertrand model demand parameters}
\usage{
bertrand_vert_calibrate(param, own_down, price, shares, cost, price_w)
}
\arguments{
\item{param}{Price coefficient, alpha}

\item{own_down}{Ownership matrix for downstream firms}

\item{price}{Downstream prices}

\item{shares}{Market shares}

\item{cost}{Marginal costs of downstream firm}

\item{price_w}{Upstream or wholesale prices, treated as costs by downstream
firms}
}
\value{
Objective function value
}
\description{
Calibrate Bertrand model demand parameters
}
\details{
This function can be used to calibrate the demand parameters of a
Bertrand model of competition that is the downstream market of a vertical
supply chain. Wholesale prices are treated as costs by the downstream firms.
Assumes logit demand.
}
