% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversion_calc.R
\name{diversion_calc}
\alias{diversion_calc}
\title{Market share calculations}
\usage{
diversion_calc(
  price,
  alpha,
  delta,
  nest_allocation = NA,
  mu = NA,
  marginal = FALSE,
  outsideOption = TRUE
)
}
\arguments{
\item{price}{Price}

\item{alpha}{Price coefficient}

\item{delta}{Mean values}

\item{nest_allocation}{For generalized nested logit demand, a J-by-K matrix
where each element (j,k) designates the membership of good j in nest k. Rows
should sum to 1.}

\item{mu}{Nesting parameters for each nest}

\item{marginal}{logical; if True, diversions are calculated as diversions from
a small price increase. If False, diversions are calculated as second choice
diversions.}

\item{outsideOption}{logical; whether to include an outside option in choice
set. Default is TRUE.}
}
\value{
Returns vector of choice probabilities for each good
}
\description{
Calculates choice probabilities based on logit or generalized nested logit
demand parameters
}
\details{
This function calculates choice probabilities based on demand
parameters for a logit or generalized nested logit demand system
}
\examples{
TO BE ADDED.

}
