% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_share.R
\name{match_share}
\alias{match_share}
\title{Market share calculations}
\usage{
match_share(delta, price, alpha, nest_allocation, mu, shares_obs)
}
\arguments{
\item{delta}{Mean values}

\item{price}{Price}

\item{alpha}{Price coefficient}

\item{nest_allocation}{For generalized nested logit demand, a J-by-K matrix
where each element (j,k) designates the membership of good j in nest k. Rows
should sum to 1.}

\item{mu}{Nesting parameters for each nest}
}
\value{
Returns vector of difference between predicted shares and observed
shares
}
\description{
Calculates choice probabilities based on logit or generalized nested logit
demand parameters
}
\details{
This function calculates the difference between model predicted
choice probabilities and observed market shares.
}
\examples{
TO BE ADDED.

}
