% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_share_logit.R
\name{match_share_logit}
\alias{match_share_logit}
\title{Market share calculations}
\usage{
match_share_logit(delta, shares, alpha, price)
}
\arguments{
\item{delta}{Mean values}

\item{shares}{Observed shares to match}

\item{alpha}{Price coefficient}

\item{price}{Price}
}
\value{
Returns vector of difference between predicted shares and observed
shares
}
\description{
Calculates choice probabilities based on logit or generalized nested logit
demand parameters
}
\details{
This function calculates the difference between model predicted
choice probabilities and observed market shares.
}
\examples{
TO BE ADDED.

}
