% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_calc.R
\name{share_calc}
\alias{share_calc}
\title{Market share calculations}
\usage{
share_calc(
  price,
  delta,
  alpha,
  a_jk = NA,
  B = NA,
  mu = NA,
  returnLogsum = FALSE
)
}
\arguments{
\item{price}{Price}

\item{delta}{Mean values}

\item{alpha}{Price coefficient}

\item{a_jk}{For generalized nested logit demand}

\item{B}{For generalized nested logit demand}

\item{mu}{Nesting parameters for each nest}

\item{returnLogsum}{logical; whether to return the denominator of the choice
probabilities (also known as the log-sum term). Defaults to FALSE, in which
case the function returns a vector with each product's choice probability.}
}
\value{
Returns vector of choice probabilities for each good
}
\description{
Calculates choice probabilities based on logit or generalized nested logit
demand parameters
}
\details{
This function calculates choice probabilities based on demand
parameters for a logit or generalized nested logit demand system
}
\examples{
TO BE ADDED.

}
