% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_bargain_foc.R
\name{ssa_bargain_foc}
\alias{ssa_bargain_foc}
\title{Second score bargaining first-order conditions}
\usage{
ssa_bargain_foc(price, own, alpha, delta, cost, lambda, includeMUI = TRUE)
}
\arguments{
\item{price}{Price}

\item{own}{Ownership matrix}

\item{alpha}{Price coefficient}

\item{delta}{Mean values}

\item{cost}{Marginal costs for each product}

\item{lambda}{Bargaining power of the buyer}

\item{includeMUI}{logical; TO BE ADDED}
}
\value{
The first-order conditions
}
\description{
Second score bargaining first-order conditions
}
\details{
This function calculate the first-order conditions from a bargaining
model that nests the second score auction.
}
\examples{
TO BE ADDED.

}
