% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_calibrate.R
\name{ssa_calibrate}
\alias{ssa_calibrate}
\title{Second score auction calibration}
\usage{
ssa_calibrate(param, own, price, share, cost, weight)
}
\arguments{
\item{param}{The price coefficient alpha to be calibrated}

\item{own}{Ownership matrix}

\item{price}{Price}

\item{share}{Share}

\item{cost}{Marginal costs for each product}

\item{weight}{Weighting matrix}
}
\value{
The first-order conditions
}
\description{
Second score auction calibration
}
\details{
This function calculate the first-order conditions from a second
score auction model of competition
}
\examples{
own_pre = diag(3)
p0 <- c(.05, .34, .33)
share1 <- c( 0.31, 0.27, 0.25)
c_j <- c(.05,.31,.30)
wt_matrix <- diag(c(1,1,1))

ssa_calibrate(param = -1,own = own_pre,price=p0,share=share1,cost=c_j,
              weight = wt_matrix)

}
