% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa_calibrate_gnl.R
\name{ssa_calibrate_gnl}
\alias{ssa_calibrate_gnl}
\title{Calibrate second score auction model with GNL demand}
\usage{
ssa_calibrate_gnl(
  param,
  own,
  price,
  shares,
  cost,
  weight,
  nest_allocation,
  mu_constraint_matrix = NA
)
}
\arguments{
\item{param}{Vector of demand parameters (alpha,mu)}

\item{own}{Ownership matrix}

\item{price}{Observed prices}

\item{shares}{Observed market shares}

\item{cost}{Marginal costs for each product}

\item{weight}{Vector of length four with weights given to prices, shares,
diversions, and costs, respectively. Default is c(1,1,1,1).}

\item{nest_allocation}{For generalized nested logit demand, a J-by-K matrix
where each element (j,k) designates the membership of good j in nest k. Rows
should sum to 1.}

\item{mu_constraint_matrix}{is a (K-by-K') matrix indicating which nesting
parameters are constrained to be equal to each other, where K is the
number of nests and K' is the number of freely varying nesting
parameters. mu_full = mu_constraint_matrix \%*\% mu_prime. Where mu_full
is a vector of length K of the nesting parameter value for each nest,
and mu_prime is a vector of length K' of parameters to be calculated.
It must be the case that K is greater than K'.}

\item{div_matrix}{A matrix of observed diversions from product in row j to
product in column k.}

\item{div_calc_marginal}{is a logical if function should match to marginal
diversions (if TRUE) or second choice diversions (if FALSE). Default
to TRUE.}

\item{returnOutcomes}{logical; should equilibrium objects be returned (mean
value parameter, prices, shares, costs) as a list.}
}
\value{
Difference between model predicted and observed values of
prices, shares, and diversions.
}
\description{
Calibrate second score auction model with GNL demand
}
\details{
This function calibrates a second score auction model with generalized nested
logit (GNL) demand
}
\examples{
TO BE ADDED.

}
