% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbargain_calibrate.R
\name{ssbargain_calibrate}
\alias{ssbargain_calibrate}
\title{Second score bargaining first-order conditions}
\usage{
ssbargain_calibrate(
  param,
  own,
  price,
  shares,
  cost,
  weight,
  lambda,
  includeMUI = TRUE
)
}
\arguments{
\item{param}{Price coefficient alpha and mean values delta, parameters to
calibrate}

\item{own}{Ownership matrix}

\item{price}{Price}

\item{shares}{Observed market shares}

\item{cost}{Marginal costs for each product}

\item{weight}{Weighting matrix}

\item{lambda}{Bargaining power of the buyer}

\item{includeMUI}{logical; whether to include marginal utility of income
in buyer's payoff, thereby translating dollars to utility. Default is True,
interpreted as buyer maximizing utility. Setting equal to False would have
interpretation that buyer maximizes profits.}
}
\value{
The first-order conditions
}
\description{
Second score bargaining first-order conditions
}
\details{
This function calculate the first-order conditions from a bargaining
model that nests the second score auction.
}
\examples{
alpha  <- -0.9
delta <- c(.81,.93,.82)
own_pre = diag(3)
p0 <- c(.05, .34, .33)
c_j <- c(.05,.31,.30)
wt_matrix <- diag(c(1,1,1,1000,1000,1000))
share1 <- c( 0.31, 0.27, 0.25)

ssbargain_calibrate(param = c(alpha,delta),own = own_pre,price = p0,
                    shares = share1, cost = c_j, weight = wt_matrix,
                    lambda = 0.5)

}
