% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbargain_foc.R
\name{ssbargain_foc}
\alias{ssbargain_foc}
\title{Second score bargaining first-order conditions}
\usage{
ssbargain_foc(price, own, alpha, delta, cost, lambda, includeMUI = TRUE)
}
\arguments{
\item{price}{Price}

\item{own}{Ownership matrix}

\item{alpha}{Price coefficient}

\item{delta}{Mean values}

\item{cost}{Marginal costs for each product}

\item{lambda}{Bargaining power of the buyer}

\item{includeMUI}{logical; TO BE ADDED}
}
\value{
The first-order conditions
}
\description{
Second score bargaining first-order conditions
}
\details{
This function calculate the first-order conditions from a bargaining
model that nests the second score auction.
}
\examples{
alpha  <- -0.9
delta <- c(.81,.93,.82)
own_pre = diag(3)
p0 <- c(.05, .34, .33)
c_j <- c(.05,.31,.30)

ssbargain_foc(price = p0,own = own_pre,alpha=alpha,delta=delta,cost=c_j,
              lambda = 0.5)

}
