% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bargain_calibrate.R
\name{bargain_calibrate}
\alias{bargain_calibrate}
\title{Nash bargaining calibration}
\usage{
bargain_calibrate(
  param,
  own,
  price,
  shares,
  cost,
  weight,
  lambda,
  includeMUI = TRUE
)
}
\arguments{
\item{param}{Price coefficient alpha and mean values delta, parameters to
calibrate}

\item{own}{Ownership matrix}

\item{price}{Price}

\item{shares}{Observed market shares}

\item{cost}{Marginal costs for each product}

\item{weight}{Weighting matrix}

\item{lambda}{Bargaining power of the buyer}

\item{includeMUI}{logical; whether to include marginal utility of income
in buyer's payoff, thereby translating dollars to utility. Default is True,
interpreted as buyer maximizing utility. Setting equal to False would have
interpretation that buyer maximizes profts.}
}
\value{
The first-order conditions
}
\description{
Nash bargaining calibration
}
\details{
This function calculate the first-order conditions from a Bertrand
price-setting model of competition
}
\examples{
TO BE ADDED.

}
