% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bargain_foc.R
\name{bargain_foc}
\alias{bargain_foc}
\title{Nash bargaining first-order conditions}
\usage{
bargain_foc(price, own, alpha, delta, cost, lambda, includeMUI = TRUE)
}
\arguments{
\item{price}{Price}

\item{own}{Ownership matrix}

\item{alpha}{Price coefficient}

\item{delta}{Mean values}

\item{cost}{Marginal costs for each product}

\item{lambda}{Bargaining power of the buyer}

\item{includeMUI}{logical; TO BE ADDED}
}
\value{
The first-order conditions
}
\description{
Nash bargaining first-order conditions
}
\details{
This function calculate the first-order conditions from a Bertrand
price-setting model of competition
}
\examples{
TO BE ADDED.

}
