% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bargain_foc_vert_sim.R
\name{bargain_foc_vert_sim}
\alias{bargain_foc_vert_sim}
\title{Nash Product FOCs for use in vertical model with simultaneous timing}
\usage{
bargain_foc_vert_sim(
  price_w,
  own_down,
  own_up,
  alpha,
  delta,
  cost_w,
  cost_r,
  lambda,
  price_r,
  sumFOC = TRUE
)
}
\arguments{
\item{price_w}{Upstream or wholesale prices}

\item{own_down}{Ownership matrix for downstream firms}

\item{own_up}{Ownership matrix for upstream firms}

\item{alpha}{Price coefficient}

\item{delta}{Mean values}

\item{cost_w}{Marginal costs for upstream firm for each product}

\item{cost_r}{Marginal costs for downstream firm for each product}

\item{lambda}{Bargaining power of the buyer/retailer}

\item{price_r}{Retail prices starting values}

\item{sumFOC}{logical; if true, returns the sum of first-order conditions, if
false, returns vector of FOC's for each product}
}
\value{
The first-order conditions
}
\description{
Nash Product FOCs for use in vertical model with simultaneous timing
}
\details{
This function calculate the first-order conditions from a Nash
bargaining model. For use in a vertical supply chain with simultaneous timing
}
\examples{
TO BE ADDED.

}
