% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bargain_NP_vert_seq.R
\name{bargain_NP_vert_seq_gnl}
\alias{bargain_NP_vert_seq_gnl}
\title{Nash Product for use in vertical model with sequential timing}
\usage{
bargain_NP_vert_seq_gnl(
  w_start,
  product_max,
  price_w,
  own_down,
  own_up,
  alpha,
  delta,
  cost_w,
  cost_r,
  lambda,
  p_R0,
  sigma,
  a_jk,
  B,
  mu,
  showAll = FALSE
)
}
\arguments{
\item{w_start}{Initial wholesale price for product for which to maximize Nash Product}

\item{product_max}{Index of product for which to maximize Nash Product}

\item{price_w}{Upstream or wholesale prices}

\item{own_down}{Ownership matrix for downstream firms}

\item{own_up}{Ownership matrix for upstream firms}

\item{alpha}{Price coefficient}

\item{delta}{Mean values}

\item{cost_w}{Marginal costs for upstream firm for each product}

\item{cost_r}{Marginal costs for downstream firm for each product}

\item{lambda}{Bargaining power of the buyer/retailer}

\item{p_R0}{Retail prices starting values}

\item{sigma}{Contract type; value between 0 and 1 where 0 is linear price,
1 is two-part tariff}

\item{showAll}{logical; if true, returns gains for trade for every product}

\item{maxJointProfits}{Experimental option; if true, will maximize joint
profits rather than gains from trade}
}
\value{
The first-order conditions
}
\description{
Nash Product for use in vertical model with sequential timing
}
\details{
This function calculate the Nash Product from a Nash bargaining
model. This version allows for nested logit with overlapping nests.
}
\examples{
TO BE ADDED.

}
