% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bargain_vert_sim_calibrate_gnl.R
\name{bargain_vert_sim_calibrate_gnl}
\alias{bargain_vert_sim_calibrate_gnl}
\title{Calibrate upstream Nash bargain in vertical model with simultaneous timing}
\usage{
bargain_vert_sim_calibrate_gnl(
  param,
  lambda = NA,
  cost_w = NA,
  own_down,
  own_up,
  alpha,
  delta,
  cost_r,
  price_r,
  price_w,
  a_jk = NA,
  B = NA,
  mu = NA,
  sumFOC = FALSE
)
}
\arguments{
\item{param}{vector of parameters to calibrate, will be either bargaining
parameter or demand parameters, depending on what other information is
supplied}

\item{lambda}{Bargaining power of the buyer/retailer}

\item{cost_w}{Marginal costs for upstream firm for each product}

\item{own_down}{Ownership matrix for downstream firms}

\item{own_up}{Ownership matrix for upstream firms}

\item{alpha}{Price coefficient}

\item{delta}{Mean values}

\item{cost_r}{Marginal costs for downstream firm for each product}

\item{price_r}{Retail prices}

\item{price_w}{Upstream or wholesale prices}
}
\value{
The value of objective function
}
\description{
Calibrate upstream Nash bargain in vertical model with simultaneous timing
}
\details{
This function can be used to calibrate the bargaining parameter
in a Nash bargain which is the upstream market of a vertical supply chain.
Assumes logit demand.
}
\examples{
TO BE ADDED.

}
