% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertrand_calibrate.R
\name{bertrand_calibrate}
\alias{bertrand_calibrate}
\title{Bertrand model calibration}
\usage{
bertrand_calibrate(param, own, price, shares, cost, weight)
}
\arguments{
\item{param}{Price coefficient alpha and mean values delta, parameters to
calibrate}

\item{own}{Ownership matrix}

\item{price}{Price}

\item{cost}{Marginal costs for each product}

\item{weight}{Weighting matrix}
}
\value{
The first-order conditions
}
\description{
Bertrand model calibration
}
\details{
This function calculate the first-order conditions from a Bertrand
price-setting model of competition
}
\examples{
TO BE ADDED.

}
