% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssbargain_calibrate.R
\name{ssbargain_calibrate}
\alias{ssbargain_calibrate}
\title{Second score bargaining first-order conditions}
\usage{
ssbargain_calibrate(
  param,
  own,
  price,
  shares,
  cost,
  weight,
  lambda,
  includeMUI = TRUE
)
}
\arguments{
\item{own}{Ownership matrix}

\item{price}{Price}

\item{cost}{Marginal costs for each product}

\item{lambda}{Bargaining power of the buyer}

\item{includeMUI}{logical; TO BE ADDED}

\item{alpha}{Price coefficient}

\item{delta}{Mean values}
}
\value{
The first-order conditions
}
\description{
Second score bargaining first-order conditions
}
\details{
This function calculate the first-order conditions from a bargaining
model that nests the second score auction.
}
\examples{
TO BE ADDED.

}
